Torrent Suite 3.6 release r59210

+++ Analysis/Separator/AvgKeyIncorporation.h
@@ -7,7 +7,7 @@ class AvgKeyIncorporation {
   virtual ~AvgKeyIncorporation() {}
   virtual float  *GetAvgKeySig (int region_num, int rStart, int rEnd, int cStart, int cEnd) = 0;
   virtual double  GetAvgKeySigLen () = 0;
-  virtual int     GetStart(int region_num, int rStart, int rEnd, int cStart, int cEnd) = 0;
+  virtual float   GetStart(int region_num, int rStart, int rEnd, int cStart, int cEnd) = 0;
 };
 
 #endif // AVGKEYINCORPORATION_H
